Ensures that walking away from a lever turns it off
It wouldn't be much of a challenge if you could leave a lever on all the time, so when you walk away it has to turn off. This, effectively marks the completion of the first level challenge. Now all that remains is to tidy it all up and make it playable