Adds a new method and makes jumping better
Adds a new method that allows inclusion of JS files. I know that Yabble seems to already do this, but this is synchronous and means I don't have to learn Yabble... win win considering the time constraints.

Also adds better jumping support. Previously, if the jump button was held down then at the top of the jump the player would jump again. This now doesn't happen